Refactor `ResetRobotControllerLights` to be a self-activating OpMode
With the help of a `@Hook` the ResetRobotControllerLights OpMode will start and stop itself immediately after the execution of a user OpMode. This opens up LynxModule access temporarily allowing us to send the reset command quickly.

Additionally refactored some methods such as `isOpModeRunning` to `isUserOpModeRunning` to disambiguate and include BunyipsLib-integrated OpModes as non-user OpModes. Added a new `@Hook` flag to bypass the new checks.